Allow `acm-operator/*` User Agent
https://issues.redhat.com/browse/RHCLOUD-14014

We need to support the following user agent value in 3scale/uhc-auth-proxy:
```
{code:java}
acm-operator/* cluster/*
{code}
```

Since the above value will pass the match for supported user agents being treated as `uhc-auth` in the gateway, this should _not_ require any changes there. We will however need to support this in `uhc-auth-proxy`.

This PR adds that support by allowing `acm-operator` as a valid operator prefix, and adds tests to exercise the changes.