Ensure we check the length of the User Agent before access by index
This adds a test to exploit a situation where you can cause the system to panic
and throw a 500, when the User Agent slice does not contain at least two substrings.

For instance, the following will currently cause a panic/500, since the expected
`cluster/` portion of the agent string is missing:

```
curl http://localhost:8080/api/uhc-auth-proxy/v1/ -A 'foo-operator/blah' -H 'Authorization: Bearer foo'

```

This resolves the issue by checking if ` cluster/*` is in the agent prior to
accessing the slice by index. We also remove duplication for returning the same error on the User Agent checks.